+++ middleware/app/api/main.py
@@ -1,5 +1,6 @@
 from api.routes import download, health, upload
 from api.routes.secured.main import secured_api_router
+from api.routes.webhook.main import webhook_api_router
 from fastapi import APIRouter
 
 api_router = APIRouter()
@@ -7,3 +8,4 @@ api_router.include_router(health.router, prefix="/health", tags=["Health Routes"
 api_router.include_router(upload.router, prefix="/upload", tags=["Upload Routes"])
 api_router.include_router(download.router, prefix="/download", tags=["Download Routes"])
 api_router.include_router(secured_api_router)
+api_router.include_router(webhook_api_router)
