+++ middleware/app/api/routes/secured/main.py
@@ -1,5 +1,4 @@
-from api.routes.secured.routes import (developer, feedback, scan, subscribe,
-                                       user)
+from api.routes.secured.routes import (developer, feedback, scan, subscribe)
 from fastapi import APIRouter
 
 secured_api_router = APIRouter(prefix="/secured")
@@ -9,7 +8,6 @@ secured_api_router.include_router(
 secured_api_router.include_router(
     subscribe.router, prefix="/subscribe", tags=["Subscribe Routes"]
 )
-secured_api_router.include_router(user.router, prefix="/user", tags=["User Routes"])
 secured_api_router.include_router(scan.router, prefix="/scan", tags=["Scan Routes"])
 secured_api_router.include_router(
     developer.router, prefix="/developer", tags=["Developer Routes"]
