+++ middleware/template.yaml
@@ -100,6 +100,13 @@ Resources:
               Ref: ByteShareAPIGW
             Path: /secured/{proxy+}
             Method: any
+        WebhookApi:
+          Type: Api
+          Properties:
+            RestApiId:
+              Ref: ByteShareAPIGW
+            Path: /webhook/{proxy+}
+            Method: any
   ByteShareAPIGW:
     Type: AWS::Serverless::Api
     Properties:
